+++ Pangle/PangleBannerCustomEvent.m
@@ -89,15 +89,22 @@ Banner size mapping according to the incoming size in adapter and selected size 
 - (CGSize)sizeForAdapterInfo:(CGSize)size {
     CGFloat width = size.width;
     CGFloat height = size.height;
-    CGFloat renderRatio = height * 1.0 / width;
-    
-    if (renderRatio >= [BUSize sizeBy:BUProposalSize_Banner600_500].height * 1.0 /
-        [BUSize sizeBy:BUProposalSize_Banner600_500].width) {
-        return CGSizeMake(width,
-                          width * [BUSize sizeBy:BUProposalSize_Banner600_500].height / [BUSize sizeBy:BUProposalSize_Banner600_500].width); //0.83
+    if (height >= floor(width * 450.0 / 600.0)) {
+        // Banner 600*500
+        if (width / height > 600 / 500) {
+            return CGSizeMake(height / 5 * 6, height);
+        }
+        else {
+            return CGSizeMake(width, width / 6 * 5);
+        }
     } else {
-        return CGSizeMake(width,
-                          width * [BUSize sizeBy:BUProposalSize_Banner640_100].height / [BUSize sizeBy:BUProposalSize_Banner640_100].width); //0.16
+        // Banner 600*150
+        if (width / height > 640 / 100) {
+            return CGSizeMake(height / 100 * 640, height);
+        }
+        else {
+            return CGSizeMake(width, width / 640 * 100);
+        }
     }
 }