+++ SnapAudienceNetwork/SnapAdRewardedVideoCustomEvent.m
@@ -5,7 +5,6 @@
 #if __has_include("MoPub.h")
     #import "MPLogging.h"
     #import "MPError.h"
-    #import "MPRewardedVideoError.h"
     #import "MPReward.h"
 #endif
 
@@ -131,7 +130,7 @@ static NSString *const kSAKSlotId = @"slotId";
 {
     MPLogAdEvent([MPLogEvent adDidAppearForAdapter:NSStringFromClass(self.class)], self.slotId);
     
-    [self.delegate fullscreenAdAdapterAdDidAppear:self];
+    [self.delegate fullscreenAdAdapterAdDidPresent:self];
     [self.delegate fullscreenAdAdapterDidTrackImpression:self];
 }
 
@@ -178,7 +177,7 @@ static NSString *const kSAKSlotId = @"slotId";
 - (void)rewardedAdWillAppear:(nonnull SAKRewardedAd *)ad
 {
     MPLogAdEvent([MPLogEvent adWillAppearForAdapter:NSStringFromClass(self.class)], self.slotId);
-    [self.delegate fullscreenAdAdapterAdWillAppear:self];
+    [self.delegate fullscreenAdAdapterAdWillPresent:self];
 }
 
 - (void)rewardedAdWillDisappear:(nonnull SAKRewardedAd *)ad