+++ Vungle/VungleRouter.m
@@ -5,11 +5,11 @@
 //  Copyright (c) 2015 MoPub. All rights reserved.
 //
 
+#import <VungleSDK/VungleSDK.h>
 #import <VungleSDK/VungleSDKHeaderBidding.h>
+#import <VungleSDK/VungleSDKNativeAds.h>
 #if __has_include("MoPub.h")
     #import "MPLogging.h"
-    #import "MPRewardedVideo.h"
-    #import "MPRewardedVideoError.h"
     #import "MoPub.h"
 #endif
 #import "VungleAdapterConfiguration.h"
@@ -39,7 +39,7 @@ typedef NS_ENUM(NSUInteger, SDKInitializeState) {
     SDKInitializeStateInitialized
 };
 
-@interface VungleRouter ()
+@interface VungleRouter () <VungleSDKDelegate, VungleSDKNativeAds>
 
 @property (nonatomic, copy) NSString *vungleAppID;
 @property (nonatomic) BOOL isAdPlaying;
@@ -131,6 +131,8 @@ typedef NS_ENUM(NSUInteger, SDKInitializeState) {
             NSError * error = nil;
             // Disable refresh functionality for all banners
             [[VungleSDK sharedSDK] disableBannerRefresh];
+            // Enable background downloading
+            [VungleSDK enableBackgroundDownload:YES];
             BOOL started = [[VungleSDK sharedSDK] startWithAppId:appId options:initOptions error:&error];
             if (!started && error.code == VungleSDKErrorSDKAlreadyInitializing) {
                 MPLogInfo(@"Vungle:SDK already has been initialized.");
@@ -225,7 +227,7 @@ typedef NS_ENUM(NSUInteger, SDKInitializeState) {
             [self requestAdWithCustomEventInfo:info delegate:delegate];
         }
     } else {
-        NSError *error = [NSError errorWithDomain:MoPubRewardedVideoAdsSDKDomain code:MPRewardedVideoAdErrorUnknown userInfo:nil];
+        NSError *error = [NSError errorWithDomain:MoPubRewardedAdsSDKDomain code:MPRewardedAdErrorUnknown userInfo:nil];
         [delegate vungleAdDidFailToLoad:error];
     }
 }
@@ -401,7 +403,7 @@ typedef NS_ENUM(NSUInteger, SDKInitializeState) {
             self.isAdPlaying = NO;
         }
     } else {
-        NSError *error = [NSError errorWithDomain:MoPubRewardedVideoAdsSDKDomain code:MPRewardedVideoAdErrorNoAdsAvailable userInfo:nil];
+        NSError *error = [NSError errorWithDomain:MoPubRewardedAdsSDKDomain code:MPRewardedAdErrorNoAdsAvailable userInfo:nil];
         [[self.delegatesDict objectForKey:placementId] vungleAdDidFailToPlay:error];
     }
 }