+++ Pangle/CHANGELOG.md
@@ -1,4 +1,32 @@
 ## Changelog
+   * 3.7.0.5.0
+     * This version of the adapters has been certified with Pangle 3.7.0.5 and MoPub SDK 5.17.0.
+
+   * 3.6.1.7.0
+     * This version of the adapters has been certified with Pangle 3.6.1.7 and MoPub SDK 5.17.0.
+
+   * 3.6.1.5.0
+     * Optimize banner size mapping.
+     * This version of the adapters has been certified with Pangle 3.6.1.5 and MoPub SDK 5.17.0.
+
+   * 3.5.1.2.1
+     * Add `fullscreenAdAdapterAdWillPresent` and `fullscreenAdAdapterAdDidPresent` to notify publishers of the fullscreen ad show event. Remove `fullscreenAdAdapterAdWillAppear` and  `fullscreenAdAdapterAdDidAppear` as they are now deprecated by the MoPub iOS SDK.
+     * Publishers must use v5.17.0 of the MoPub SDK at the minimum.
+
+   * 3.5.1.2.0
+     * This version of the adapters has been certified with Pangle 3.5.1.2 and MoPub SDK 5.16.2.
+
+   * 3.5.1.0.1
+     * Remove deprecated usage of Rewarded Video APIs.
+
+   * 3.5.1.0.0
+     * This version of the adapters has been certified with Pangle 3.5.1.0 and MoPub SDK 5.16.2.
+
+   * 3.4.4.3.0
+     * Change the SDK pod dependency from `Bytedance-UnionAD` to `Ads-Global`. No action required for publishers.
+     * Pangle iOS SDK was separated into Pangle China SDK and Pangle non-China SDK from the version 3.4.0.0. MoPub will only certify Pangle non-China SDK start from v3.4.4.3.
+     * This version of the adapters has been certified with Pangle 3.4.4.3 and MoPub SDK 5.16.2.
+
    * 3.3.6.2.2
      * Add support for `fullscreenAdAdapterAdWillDismiss` when a fullscreen creative is about to close. Publishers must use v5.16.1 of the MoPub SDK at the minimum.